Update from float to int fields in book model

+++ library_project/library_app/models.py
@@ -9,8 +9,8 @@ class Author(models.Model):
 
 class Book(models.Model):
     name = models.CharField(max_length=100)
-    edition = models.FloatField()
-    publication_year = models.FloatField()
+    edition = models.IntegerField()
+    publication_year = models.IntegerField()
     authors = models.ManyToManyField(Author)
 
     def __str__(self):
